+++ samples/spring-boot-demo/src/main/java/io/github/rabitem/outcomemetrics/samples/spring/SpringDemoApplication.java
@@ -1,0 +1,12 @@
+package io.github.rabitem.outcomemetrics.samples.spring;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringDemoApplication {
+
+    public static void main(final String[] args) {
+        SpringApplication.run(SpringDemoApplication.class, args);
+    }
+}